requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
mlflow==2.5.0
dvc==3.15.0
optuna==3.12.0
plotly==5.15.0
pyyaml==6.0
pytest==7.4.0
black==23.7.0
isort==5.12.0
flake8==6.0.0
xgboost==1.7.6
requests==2.31.0
